requirements.txt
fastapi
pydantic
PyYAML
selenium
SQLAlchemy
uvicorn
celery
redis
psycopg2-binary